# src/main/java/com/example/ecommerce/ecommerce/_swagger/swagger.yaml
openapi: 3.0.3
info:
  title: Mini E-Commerce API
  version: 1.0.0
  description: |
     E-Commerce backend.
servers:
  - url: http://localhost:8080/api/v1
...
paths:
...
  /products/{id}:
    get:
      tags: [Products]
      summary: Get product by ID
      parameters:
        - $ref: '#/components/parameters/IdPath'
      responses:
        '200':
          description: Product
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductVTO'
    put:
      tags: [Products]
      summary: Update product
      parameters:
        - $ref: '#/components/parameters/IdPath'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProductUpdateRequestDTO'
      responses:
        '200':
          description: Updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductVTO'
    delete:
      tags: [Products]
      summary: Delete product
      parameters:
        - $ref: '#/components/parameters/IdPath'
      responses:
        '204':
          description: Deleted
...
components:
  parameters:
    IdPath:
      name: id
      in: path
      required: true
      schema: { type: integer, format: int64 }
  schemas:
...
    ProductUpdateRequestDTO:
      type: object
      properties:
        name: { type: string }
        description: { type: string }
        price: { type: number }
        stock: { type: integer }
    ProductVTO:
      type: object
      properties:
        id: { type: integer }
        name: { type: string }
        description: { type: string }
        price: { type: number }
        stock: { type: integer }
...